Move `processStatusUpdate` tests to common test.
This method is shared by both deployment modes, so test it in the common
test file.
